+++ tests/diagnostics/mod.rs
@@ -1,4 +1,5 @@
 mod basic;
+mod bloat;
 mod fix;
 mod indexes;
 mod locks;